server/requirements.txt
APScheduler==3.10.1
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.6.0
attrs==21.2.0
autopep8==2.0.2
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
ccsm==0.9.14.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.0.3
cmake==3.26.3
colorama==0.4.4
command-not-found==0.3
compizconfig-python==0.9.14.1
configobj==5.0.6
contourpy==1.0.7
cryptography==40.0.2
cupshelpers==1.0
cvzone==1.5.6
cycler==0.11.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
dj-rest-auth==4.0.0
Django==3.2.12
django-allauth==0.52.0
django-apscheduler==0.6.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
eyeD3==0.8.10
filelock==3.6.0
filterpy==1.4.5
flake8==6.0.0
fonttools==4.39.3
grpcio==1.30.2
gyp==0.1
httpie==3.2.2
httplib2==0.20.2
idna==3.4
ifaddr==0.1.7
imageio==2.28.0
IMDbPY==2021.4.18
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.2
lit==16.0.2
louis==3.20.0
mailtrap==2.0.1
Mako==1.1.3
markdown-it-py==2.2.0
MarkupSafe==2.0.1
matplotlib==3.7.1
mccabe==0.7.0
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
netaddr==0.8.0
netifaces==0.11.0
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onboard==1.4.1
opencv-python==4.7.0.72
packaging==21.3
pandas==2.0.0
pdf2image==1.16.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
pyflakes==3.0.1
Pygments==2.15.1
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyrsistent==0.18.1
PySocks==1.7.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-magic==0.4.24
python-xapp==2.2.2
python-xlib==0.29
python3-openid==3.2.0
pytz==2023.3
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.3.5
scikit-image==0.20.0
scipy==1.10.1
seaborn==0.12.2
SecretStorage==3.3.1
sentry-sdk==1.21.0
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.4
sympy==1.11.1
systemd-python==234
texttable==1.6.4
thop==0.1.1.post2209072238
tifffile==2023.4.12
tldextract==3.1.2
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ultralytics==8.0.88
Unidecode==1.3.3
urllib3==1.26.15
wadllib==1.3.6
websocket-client==1.2.3
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
